introduce-invariant-checker-pre-move: Move create_scheduler_watchdog from runtime_checker mixin to scheduler.py
Pre-prep for the ``introduce-invariant-checker`` mech split.

Cut ``create_scheduler_watchdog`` (module-level free function) from
``scheduler_runtime_checker_mixin.py`` and paste verbatim just before
``class Scheduler(`` in ``scheduler.py``. Wrap the ``scheduler: Scheduler``
annotation in a string forward reference (``scheduler: "Scheduler"``)
because ``Scheduler`` is defined later in the same module.

Imports updated:
- ``scheduler.py``: drop ``create_scheduler_watchdog`` from the grouped
  ``from sglang.srt.managers.scheduler_runtime_checker_mixin import (...)``
  block; add ``from sglang.srt.utils.watchdog import WatchdogRaw``.
- ``scheduler_runtime_checker_mixin.py``: drop the now-unused
  ``from sglang.srt.utils.watchdog import WatchdogRaw`` import.

Body otherwise byte-identical (``git --color-moved=dimmed-zebra`` shows the
moved function as a single hunk). This is a pure mechanical relocation of
an already-existing free function (per MECH_COMMIT_SPLIT.md "何时不拆"
single-commit exception).

Refactor chain ID: introduce-invariant-checker-pre-move
